requirements.txt
# ── Web framework ─────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# ── Data validation ────────────────────────────────────────────────────────────
pydantic>=2.7.0
pydantic-settings>=2.2.0

# ── Vector database ────────────────────────────────────────────────────────────
chromadb>=0.5.0

# ── Embeddings ─────────────────────────────────────────────────────────────────
sentence-transformers>=2.7.0

# ── PDF parsing (PyMuPDF — no system Poppler needed) ──────────────────────────
PyMuPDF>=1.24.0

# ── DOCX parsing ──────────────────────────────────────────────────────────────
python-docx>=1.1.0

# ── OCR (optional — requires Tesseract binary installed separately) ────────────
pytesseract>=0.3.10
Pillow>=10.0.0

# ── LLM ───────────────────────────────────────────────────────────────────────
groq>=0.9.0

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
numpy>=1.24.0
aiofiles>=23.2.0
